requirements.txt
pandas
python-dotenv

#openai
langchain
#langchain-openai
langchain-google-genai
#langchain-chroma
langchain-community
langsmith

jq
#pyjq

streamlit
faiss-cpu
beautifulsoup4
tiktoken
pypdf
PyPDF2
lxml

DateTime

typing
requests
ipykernel
numexpr
duckduckgo-search
Wikipedia
